feat(products):completed design and api connection feat(warehouses):completed design and api connection feat(stock levels):completed design and api connection

+++ server/inventory-service/src/warehouses/warehouse.entity.ts
@@ -22,7 +22,7 @@ export class Warehouse {
   @Column({ nullable: true })
   user_id: string;
 
-  @Column('numeric')
+  @Column('integer')
   capacity: number;
 
   @CreateDateColumn()
